очень большой коммит, исправил модели транзакций, добавил необходимые зависимости в dagger, добавил репозитории для поддержания работы приложения оффлайн и онлайн

+++ app/src/main/java/com/example/bankapp/data/local/dao/SyncOperationDao.kt
@@ -27,7 +27,7 @@ interface SyncOperationDao {
     LIMIT 1
 """
     )
-    suspend fun findExistingOperation(type: OperationType, targetId: Long): SyncOperationEntity?
+    suspend fun findExistingOperation(type: OperationType, targetId: Int): SyncOperationEntity?
 
     @Query("""
     DELETE FROM sync_operations
